correcting events info
